Mark a module dependency as exported when a module is exporting a package from a module it depends on.
This can happen for example when you'd have an api and impl project where
the impl project is exporting the api. When a 3rd project would now have
a dependency on the impl and not the api the api packages were not visible.
Marking the dependency as exported resolves this.